fix bug with signout
