+++ src/app/components/comments/comment-list/comment-list.component.scss
@@ -1,3 +1,6 @@
+button {
+  margin: 3px;
+}
 @import "../../../../sass/vars.scss";
 
 h1, h2, h3, h4, h5, h6 {
